+++ .env.example
@@ -3,7 +3,7 @@ DATABASE_URL=postgresql://postgres:postgres@localhost:${DATABASE_PORT}
 
 SHOPIFY_API_KEY=f7208c484c27f835e1c5a6dc9636045c
 
-SHORTAGE_API_ROOT=https://app.shortage.global/api/
-SHORTAGE_ROOT=https://shortage.global/
+VITE_SHORTAGE_API_ROOT=https://app.shortage.global/api/
+VITE_SHORTAGE_ROOT=https://shortage.global/
 
 SHORTAGE_AUTHORIZATION_KEY=
